testing/web-platform/tests/dom/nodes/Document-contentType/contentType/contenttype_mimeheader_03.html

<!DOCTYPE html>
<title>Custom document.contentType === 'text/html' when no Content-Type header is provided in HTTP response</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id=log></div>
<iframe style="display:none"></iframe>
<script>
async_test(function() {
  var iframe = document.getElementsByTagName('iframe')[0];
  iframe.addEventListener('load', this.step_func_done(function() {
    assert_equals(iframe.contentDocument.contentType, "text/html");
  }), false);
  iframe.src = "../support/contenttype_setter.py?removeContentType=true";
});
</script>
